Revert "Much refactor"
