+++ app/views/ui/dialogs.jade
@@ -1,9 +1,9 @@
 #dialog-box-shield
 
-#about.dialog-box.about-dialog
+#about.dialog-box
   button.close ×
   include ./dialogs/about
 
-#save-as-image-dialog.dialog-box.save-as-image-dialog
+#save-as-image-dialog.dialog-box
   button.close ×
   include ./dialogs/save-as-image
