+++ app/views/ui/dialogs/about.jade
@@ -1,8 +1,8 @@
-.about-dialog-left
+.left
   h1(data-i18n='dialogs.about.heading').
     About Streetmix.
 
-  .about-dialog-description(data-i18n='dialogs.about.description').
+  p.description(data-i18n='dialogs.about.description').
     Design, remix, and share your street.
     Add bike paths, widen sidewalks or traffic lanes, learn
     how all of this can impact your community.
@@ -15,12 +15,12 @@
       a(href='https://github.com/codeforamerica/streetmix/', target='_blank', data-i18n='dialogs.about.view-source').
         View source code
 
-.about-dialog-right
+.right
   p.
     A side project by
     <a target='_blank' href='http://codeforamerica.org'>Code for America</a> 2013 fellows:
 
-  ul.about-dialog-team
+  ul.team
     li.
       <a target='_blank' href='http://twitter.com/anselmbradford'><div class='avatar' postpone='true' userId='anselmbradford'></div>Anselm Bradford</a> · media production
     li.
