+++ assets/css/_info-bubble.scss
@@ -154,6 +154,104 @@
     }
   }
 
+  .description-prompt,
+  .description-close {
+    border-top: 1px solid $ui-colour;
+    text-rendering: optimizeLegibility;
+
+    white-space: normal;
+    text-align: center;
+
+    margin-top: 1px;
+    padding: 5px;
+    padding-top: 3px;
+    font-size: 13px;
+
+    user-select: none;
+
+    color: darken($ui-colour, 20%);
+    cursor: pointer;
+
+    &:hover {
+      color: darken($ui-colour, 40%);
+      background: mix($ui-colour, white, 50%);
+    }
+  }
+
+  .description-close {
+    position: absolute;
+    left: 0;
+    right: 0;
+    bottom: 0;
+
+    pointer-events: none;
+  }
+
+  .description-canvas {
+    position: absolute;
+    left: -50px;
+    top: -4px;
+    right: -50px;
+
+    opacity: 0;
+    pointer-events: none;
+
+    border-top: 3px solid darken($ui-colour, 20%);
+    border-bottom: 3px solid darken($ui-colour, 20%);
+
+    transform-origin: 50% 0;
+    transform: rotateX(20deg) translateY(100px);
+    transition: opacity 250ms, transform 250ms;
+
+    background: $info-bubble-background;
+  }
+
+  .description {
+    font-size: 16px;
+    padding: 12px;
+    line-height: 23px;
+    font-weight: 300;
+
+    overflow-y: auto;
+    overflow-x: hidden;
+    touch-action: pan-y;
+
+    position: absolute;
+    left: 0;
+    top: 0;
+    right: 0;
+    bottom: 25px;
+
+    p {
+      margin: 10px 0;
+
+      &:first-child {
+        margin-top: 0;
+      }
+
+      &.lede {
+        font-size: 150%;
+        font-weight: 300;
+        line-height: 1.25;
+      }
+    }
+
+    footer {
+      text-align: right;
+      color: rgb(160, 160, 160);
+      font-size: 11px;
+      line-height: 14px;
+    }
+
+    img {
+      //width: 300px;
+      margin: 0 auto;
+      display: block;
+      max-width: 100%;
+      outline: 1px solid darken($ui-colour, 20%);
+    }
+  }
+
   .warnings {
     margin-top: 1px;
     font-size: 13px;
@@ -299,101 +397,4 @@
   .description {
     pointer-events: auto;
   }
-}
-
-.description-prompt,
-.description-close {
-  border-top: 1px solid $ui-colour;
-  text-rendering: optimizeLegibility;
-
-  white-space: normal;
-  text-align: center;
-
-  margin-top: 1px;
-  padding: 10px 5px;
-  font-size: 13px;
-
-  user-select: none;
-
-  color: darken($ui-colour, 20%);
-  cursor: pointer;
-
-  &:hover {
-    color: darken($ui-colour, 40%);
-    background: mix($ui-colour, white, 50%);
-  }
-}
-
-.description-close {
-  position: absolute;
-  left: 0;
-  right: 0;
-  bottom: 0;
-
-  pointer-events: none;
-}
-
-.description-canvas {
-  position: absolute;
-  left: -50px;
-  top: -50px;
-  right: -50px;
-
-  opacity: 0;
-  pointer-events: none;
-
-  border-top: 3px solid darken($ui-colour, 20%);
-  border-bottom: 3px solid darken($ui-colour, 20%);
-
-  transform-origin: 50% 0;
-  transform: rotateX(20deg) translateY(100px);
-  transition: opacity 250ms, transform 250ms;
-
-  background: $info-bubble-background;
-}
-
-.description {
-  font-size: 16px;
-  padding: 12px;
-  line-height: 23px;
-  font-weight: 300;
-
-  overflow-y: auto;
-  overflow-x: hidden;
-  touch-action: pan-y;
-
-  position: absolute;
-  left: 0;
-  top: 0;
-  right: 0;
-  bottom: 37px;
-
-  p {
-    margin: 10px 0;
-
-    &:first-child {
-      margin-top: 0;
-    }
-  }
-
-  footer {
-    text-align: right;
-    color: rgb(160, 160, 160);
-    font-size: 11px;
-    line-height: 14px;
-  }
-
-  img {
-    //width: 300px;
-    margin: 0 auto;
-    display: block;
-    max-width: 100%;
-    outline: 1px solid darken($ui-colour, 20%);
-  }
-}
-
-.description-lede {
-  font-size: 150%;
-  font-weight: 300;
-  line-height: 1.25;
 }
