+++ assets/js/app/event_listeners.js
@@ -2,6 +2,8 @@ function _addEventListeners () {
   document.querySelector('#street-scroll-indicator-left').addEventListener('pointerdown', _onStreetLeftScrollClick)
   document.querySelector('#street-scroll-indicator-right').addEventListener('pointerdown', _onStreetRightScrollClick)
 
+  document.querySelector('#welcome .close').addEventListener('pointerdown', _hideWelcome)
+
   document.querySelector('#street-section-outer').addEventListener('scroll', _onStreetSectionScroll)
 
   document.querySelector('#street-section-left-building').addEventListener('pointerenter', _onBuildingMouseEnter)
