+++ assets/js/segments/drag_and_drop.js
@@ -1,3 +1,5 @@
+var TRACK_LABEL_DRAGGING = 'Dragging'
+
 var DRAG_OFFSET_Y_PALETTE = -340 - 150
 // Disable Y offset on segments when touch dragging
 //var DRAG_OFFSET_Y_TOUCH_PALETTE = -100
@@ -705,7 +707,8 @@ function _handleSegmentMoveEnd (event) {
       removeElFromDOM(draggingMove.originalEl)
     }
 
-    trackEvent('INTERACTION', 'REMOVE_SEGMENT', 'DRAGGING', null, true)
+    trackEvent('Interaction', TRACK_ACTION_REMOVE_SEGMENT,
+      TRACK_LABEL_DRAGGING, null, true)
   } else if (draggingMove.segmentBeforeEl || draggingMove.segmentAfterEl || (street.segments.length == 0)) {
     var smartDrop = _doDropHeuristics(draggingMove.originalType,
       draggingMove.originalVariantString, draggingMove.originalWidth)
