+++ assets/js/streets/data_model.js
@@ -250,10 +250,10 @@ function _saveStreetToServerIfNecessary () {
       // console.log('not incrementing editCount since null')
     }
     _setUpdateTimeToNow()
+    _hideWelcome()
 
-    // Some parts of the UI need to know this happened to respond to it
-    // TODO: figure out appropriate event name
-    window.dispatchEvent(new CustomEvent('stmx:save_street'))
+    // As per issue #306.
+    hideStatusMessage()
 
     _updateStreetMetadata(street)
