+++ assets/scripts/app/event_tracking.js
@@ -4,44 +4,12 @@
 // onlyFirstTime passed to eventTracking.track()
 let alreadyTracked = []
 
-// These constants define what are valid actions, labels and categories for Streetmix.
-// They're stored as strings so that they can be passed around instead of needing to
-// be stored as global variables or have to be exported from different sources.
-// passed as variables, which would need to be global (and harder to manage).
-// For a similar system see the Redux state store which similarly uses strings-as-
-// constants to define action types.
-// Strings are not namespaced to keep trackEvent() calls shorter.
-const TRACK_CATEGORY = {
-  'INTERACTION': 'Interaction',
-  'EVENT': 'Event',
-  'ERROR': 'Error',
-  'SYSTEM': 'System',
-  'SHARING': 'Sharing'
-}
-
-const TRACK_ACTION = {
-  'TOUCH_CAPABLE': 'Touch capability detected',
-  'OPEN_GALLERY': 'Open gallery',
-  'UNDO': 'Undo',
-  'CHANGE_WIDTH': 'Change width',
-  'REMOVE_SEGMENT': 'Remove segment',
-  'LEARN_MORE': 'Learn more about segment',
-  'FACEBOOK': 'Facebook',
-  'TWITTER': 'Twitter',
-  'ERROR_15A': 'Error 15A (sign in API failure)',
-  'ERROR_RM1': 'Error RM1 (auth 401 failure on load)',
-  'ERROR_RM2': 'Error RM2 (auth 401 failure mid-flight)',
-  'ERROR_GEOLOCATION_TIMEOUT': 'Geolocation timeout',
-  'STREET_MODIFIED_ELSEWHERE': 'Street modified elsewhere'
-}
-
-const TRACK_LABEL = {
-  'BUTTON': 'Button',
-  'DRAGGING': 'Dragging',
-  'INPUT_FIELD': 'Input field',
-  'INCREMENT_BUTTON': 'Increment button',
-  'KEYBOARD': 'Keyboard'
-}
+// These constants define what are valid categories for Streetmix
+// const TRACK_CATEGORY_INTERACTION = 'Interaction'
+// const TRACK_CATEGORY_EVENT = 'Event'
+// const TRACK_CATEGORY_ERROR = 'Error'
+// const TRACK_CATEGORY_SYSTEM = 'System'
+// const TRACK_CATEGORY_SHARING = 'Sharing'
 
 /**
  * Tracks an event to Google Analytics
@@ -58,13 +26,9 @@ export function trackEvent (category, action, label, value, onlyFirstTime) {
     return
   }
 
-  const trackCategory = TRACK_CATEGORY[category] || category
-  const trackAction = TRACK_CATEGORY[action] || action
-  const trackLabel = TRACK_CATEGORY[label] || label
-
   // If this should only be tracked once, do this
   if (onlyFirstTime) {
-    const id = category + '|' + action + '|' + label
+    var id = category + '|' + action + '|' + label
 
     // If it has already tracked, exit this function
     if (alreadyTracked[id]) {
