+++ assets/scripts/app/load_resources.js
@@ -4,10 +4,32 @@
  * Loads images, etc and tracks progress. (WIP)
  * TODO: Rely on Promises to resolve progress
  */
-/* global _checkIfEverythingIsLoaded, Image */
+/* global _checkIfEverythingIsLoaded */
+const SVGStagingEl = document.getElementById('svg')
+
+export let iconsSVG = window.fetch('/assets/images/icons.svg')
+  .then(function (response) {
+    return response.text()
+  })
+  .then(function (response) {
+    SVGStagingEl.innerHTML = response
+    return response
+  })
+  .catch(function (error) {
+    console.log('doh', error)
+  })
+
+export function hideLoadingScreen () {
+  // NOTE:
+  // This function might be called on very old browsers. Please make
+  // sure not to use modern faculties.
+
+  document.getElementById('loading').className += ' hidden'
+}
 
 // Image tileset loading
-// TODO: Deprecate in favor of inlined SVGs
+// TODO: Deprecate in favor of inlined SVGs & loading by promises
+
 const TILESET_IMAGE_VERSION = 55
 const IMAGES_TO_BE_LOADED = [
   '/images/tiles-1.png',
@@ -17,98 +39,38 @@ const IMAGES_TO_BE_LOADED = [
   '/images/sky-rear.png'
 ]
 
-const SVGS_TO_BE_LOADED = [
-  '/assets/images/icons.svg'
-]
-
-const SVGStagingEl = document.getElementById('svg')
-
 let images = []
-let loading = []
-
-// Set loading bar
-const loadingEl = document.getElementById('loading-progress')
-loadingEl.max += 5 // Legacy; this is for other things that must load
-
-// Global for legacy reasons
-window.imagesToBeLoaded = 1
-
-// Load everything
-loadImages()
-loadSVGs()
-
-// When everything is loaded...
-Promise.all(loading)
-  .then(function () {
-    // Export to window (LEGACY)
-    window.images = images
-    window.imagesToBeLoaded = 0
-
-    // Also legacy, TODO: replace with promise
-    _checkIfEverythingIsLoaded()
-  })
-
+let imagesToBeLoaded
+// Global: TEMP
+window.images = images
+window.imagesToBeLoaded = imagesToBeLoaded
 
 function loadImages () {
-  loadingEl.max += IMAGES_TO_BE_LOADED.length
+  imagesToBeLoaded = IMAGES_TO_BE_LOADED.length
 
-  for (let url of IMAGES_TO_BE_LOADED) {
-    loading.push(getImage(url + '?v' + TILESET_IMAGE_VERSION)
-      .then(function (image) {
-        // Store on the global images object, using the url as the key
-        images[url] = image
+  for (var i in IMAGES_TO_BE_LOADED) {
+    var url = IMAGES_TO_BE_LOADED[i]
 
-        loadingEl.value++
-      })
-      .catch(function (error) {
-        console.error('loading image error', error)
-      }))
+    images[url] = document.createElement('img')
+    images[url].addEventListener('load', onImageLoaded)
+    images[url].src = url + '?v' + TILESET_IMAGE_VERSION
   }
+
+  document.querySelector('#loading-progress').value = 0
+  document.querySelector('#loading-progress').max = imagesToBeLoaded + 5
 }
 
-function loadSVGs () {
-  loadingEl.max += SVGS_TO_BE_LOADED.length
+function onImageLoaded () {
+  imagesToBeLoaded--
+  document.querySelector('#loading-progress').value++
 
-  for (let url of SVGS_TO_BE_LOADED) {
-    loading.push(window.fetch(url)
-      .then(function (response) {
-        return response.text()
-      })
-      .then(function (response) {
-        SVGStagingEl.innerHTML += response
-        loadingEl.value++
-      })
-      .catch(function (error) {
-        console.error('loading svg error', error)
-      }))
-  }
+  // Export to window (TEMP)
+  // Doing this as an export one time did not work,
+  // so we export to window each time loading occurs
+  window.images = images
+  window.imagesToBeLoaded = imagesToBeLoaded
+
+  _checkIfEverythingIsLoaded()
 }
 
-/**
- * Wraps `new Image()`'s onload and onerror properties with
- * Promises. This provides a more reliable way of knowing when a
- * image element is done loading (rather than fetch) because a
- * fetch can successfully load an image file but be resolved
- * before the image element is ready. So we must resolve only after
- * the image's `onload` callback has been called.
- */
-function getImage (url) {
-  return new Promise(function (resolve, reject) {
-    var img = new Image()
-    img.onload = function () {
-        resolve(img)
-    }
-    img.onerror = function () {
-        reject('unable to load image ' + url)
-    }
-    img.src = url
-  })
-}
-
-export function hideLoadingScreen () {
-  // NOTE:
-  // This function might be called on very old browsers. Please make
-  // sure not to use modern faculties.
-
-  document.getElementById('loading').className += ' hidden'
-}
+loadImages()
